support receiving requirements directly (#4)
Moves the current `--requirements` flag to be `--requirements-path` (for files on disk) and adds a new argument `--requirements` which corresponds to receiving requirements from the command line. This can also be passed in as `VENVCACHE_REQUIREMENTS`.

The idea here is that some CI systems (e.g. Buildkite) let you set environment variables, so this provides a mechanism to run Python like:

```yaml
- label: Whatever
  id: whatever
  command: venvcache -- -m path.to.module
  env:
    VENVCACHE_REQUIREMENTS: |
      requests==1.2.3
```